requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.36.18
botocore==1.36.18
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.1
defusedxml==0.8.0rc2
dj-database-url==3.0.1
Django==5.1.6
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-guardian==2.4.0
django-redis==5.4.0
django-ses==4.4.0
django-storages==1.14.6
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
dotenv==0.9.9
flower==2.0.1
gunicorn==23.0.0
humanize==4.12.2
idna==3.10
jmespath==1.0.1
kombu==5.5.2
oauthlib==3.2.2
packaging==25.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.11.2
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
tornado==6.4.2
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
